+++ python/test/gluon/test_frontend.py
@@ -8,17 +8,13 @@ from triton.experimental import gluon
 from triton.experimental.gluon import language as ttgl
 from triton.experimental.gluon.language.nvidia import blackwell
 from triton.experimental.gluon.language.nvidia import hopper
+from triton.experimental.gluon.language.amd import cdna3
 from triton.experimental.gluon.language.nvidia.blackwell import mbarrier, tma, TensorMemoryLayout, async_copy
 from triton.experimental.gluon.nvidia.hopper import TensorDescriptor
 from triton._filecheck import filecheck_test, run_parser
 from triton.runtime.jit import MockTensor
 import triton.language as tl
 from triton.compiler.errors import CompilationError, CompileTimeAssertionFailure
-from triton._C.libtriton import ir
-from triton.compiler.code_generator import ast_to_ttir
-from triton.compiler import make_backend
-from triton.experimental.gluon._runtime import GluonASTSource
-from triton.backends.compiler import GPUTarget
 
 TARGET_PAT = re.compile('ttg.target = "[^"]*"')
 # HIP backend can add this attribute to function parameters
@@ -1324,9 +1320,9 @@ def test_auto_layout_broadcast():
 
 @gluon.jit
 def amd_mfma_layout_kernel():
-    mfma_layout: ttgl.constexpr = ttgl.AMDMFMALayout(version=4, instr_shape=[32, 32], transposed=True,
-                                                     warps_per_cta=[4, 1], tiles_per_warp=[4, 1], ctas_per_cga=[1, 1],
-                                                     cta_split_num=[1, 1], cta_order=[1, 0])
+    mfma_layout: ttgl.constexpr = cdna3.AMDMFMALayout(version=4, instr_shape=[32, 32], transposed=True,
+                                                      warps_per_cta=[4, 1], tiles_per_warp=[4, 1], ctas_per_cga=[1, 1],
+                                                      cta_split_num=[1, 1], cta_order=[1, 0])
 
     layout: ttgl.constexpr = ttgl.BlockedLayout([1, 1], [1, 32], [4, 1], [1, 0])
 
@@ -1338,30 +1334,13 @@ def amd_mfma_layout_kernel():
 def test_amd_mfma_layout(fresh_knobs):
     knobs.compilation.disable_line_info = True
 
-    context = ir.context()
-    ir.load_dialects(context)
-    options = dict(sanitize_overflow=False)
-    stub_target = GPUTarget("hip", 'gfx950', 64)
-    stub_backend = make_backend(stub_target)
-    stub_backend.load_dialects(context)
-    options = stub_backend.parse_options(options)
-    codegen_fns = stub_backend.get_codegen_implementation(options)
-    signature = {}
-    src = GluonASTSource(fn=amd_mfma_layout_kernel, signature=signature)
-    builder = ir.builder(context)
-    module = builder.create_module()
-    module.set_attr("ttg.threads-per-warp", builder.get_int32_attr(64))
-    module.set_attr("ttg.num-warps", builder.get_int32_attr(4))
-    module.set_attr("ttg.num-ctas", builder.get_int32_attr(1))
-    module = ast_to_ttir(amd_mfma_layout_kernel, src, context=context, options=options, codegen_fns=codegen_fns,
-                         module_map=dict(), module=module)
-    assert module.verify()
+    h = amd_mfma_layout_kernel.warmup(sanitize_overflow=False, grid=(1, ))
 
     expecttest.assert_expected_inline(
-        module.str_nodebug(), """\
+        anonymize_ir(h.asm['source']), """\
 #blocked = #ttg.blocked<{sizePerThread = [1, 1], threadsPerWarp = [1, 64], warpsPerCTA = [4, 1], order = [1, 0]}>
 #mma = #ttg.amd_mfma<{version = 4, warpsPerCTA = [4, 1], tilesPerWarp = [4, 1], instrShape = [32, 32], isTransposed = true}>
-module attributes {"ttg.num-ctas" = 1 : i32, "ttg.num-warps" = 4 : i32, "ttg.threads-per-warp" = 64 : i32} {
+module attributes {"ttg.num-ctas" = 1 : i32, "ttg.num-warps" = 4 : i32, ttg.target = "...", "ttg.threads-per-warp" = 64 : i32} {
   tt.func public @amd_mfma_layout_kernel() attributes {noinline = false} {
     %cst = arith.constant 0.000000e+00 : f32 loc(#loc)
     %cst_0 = arith.constant dense<0.000000e+00> : tensor<128x32xf32, #blocked> loc(#loc)
